[refactor] moved expr2string and expr2cpp from clang frontend to util
