+++ src/esbmc/document_subgoals.cpp
@@ -1,4 +1,4 @@
-#include <clang-c-frontend/expr2string.h>
+#include <util/expr2string.h>
 #include <esbmc/document_subgoals.h>
 #include <fstream>
 #include <util/i2string.h>
