+++ src/goto-programs/goto_check.cpp
@@ -1,5 +1,5 @@
 #include <goto-programs/goto_check.h>
-#include <clang-c-frontend/expr2string.h>
+#include <util/expr2string.h>
 #include <util/arith_tools.h>
 #include <util/array_name.h>
 #include <util/base_type.h>
