+++ src/util/c_link.cpp
@@ -1,4 +1,4 @@
-#include <clang-c-frontend/expr2string.h>
+#include <util/expr2string.h>
 #include <unordered_set>
 #include <util/base_type.h>
 #include <util/c_link.h>
