+++ src/util/expr2string.cpp
@@ -1,4 +1,4 @@
-#include <clang-c-frontend/expr2string.h>
+#include <util/expr2string.h>
 #include <util/arith_tools.h>
 #include <util/c_misc.h>
 #include <util/c_types.h>
